requirements.txt
python-telegram-bot
flask
numpy
beautifulsoup4
jieba
wordcloud
pandas
lxml
aiohttp
